+++ src/api/resolvers/spotResolver.ts
@@ -1,8 +1,10 @@
+import {GraphQLError} from 'graphql';
 import {PalletSpot} from '../../interfaces/PalletSpot';
 import {Spot} from '../../interfaces/Spot';
 import spotModel from '../models/spotModel';
 import gapModel from '../models/gapModel';
 import rowModel from '../models/rowModel';
+import {UserIdWithToken} from '../../interfaces/User';
 
 export default {
   PalletSpot: {
@@ -35,11 +37,29 @@ export default {
     },
   },
   Mutation: {
-    createSpot: async (_parent: undefined, args: Spot) => {
+    createSpot: async (
+      _parent: undefined,
+      args: Spot,
+      user: UserIdWithToken
+    ) => {
+      if (!user.token) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'UNAUTHENTICATED'},
+        });
+      }
       const spot = new spotModel(args);
       return await spot.save();
     },
-    createSpots: async (_parent: undefined, args: any) => {
+    createSpots: async (
+      _parent: undefined,
+      args: any,
+      user: UserIdWithToken
+    ) => {
+      if (!user.token) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'UNAUTHENTICATED'},
+        });
+      }
       for (let i = 0; i < args.numberOfRows; i++) {
         const row = new rowModel({
           rowNumber: i + 1,
@@ -64,10 +84,28 @@ export default {
         }
       }
     },
-    updateSpot: async (_parent: undefined, args: Spot) => {
+    updateSpot: async (
+      _parent: undefined,
+      args: Spot,
+      user: UserIdWithToken
+    ) => {
+      if (!user.token) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'UNAUTHENTICATED'},
+        });
+      }
       return await spotModel.findByIdAndUpdate(args.id, args, {new: true});
     },
-    deleteSpot: async (_parent: undefined, args: Spot) => {
+    deleteSpot: async (
+      _parent: undefined,
+      args: Spot,
+      user: UserIdWithToken
+    ) => {
+      if (!user.token) {
+        throw new GraphQLError('Not authorized', {
+          extensions: {code: 'UNAUTHENTICATED'},
+        });
+      }
       return await spotModel.findByIdAndDelete(args.id);
     },
   },
